docker.version/requirements.txt
# Production dependencies for Docker deployment
# Python 3.10 compatible versions from main project
fastapi>=0.128.2
httpx>=0.28.1
icalendar>=6.3.2
lxml>=6.0.2
mcp>=1.2.0
pydantic>=2.12.5
python-dateutil>=2.9.0.post0
python-dotenv>=1.2.1
uvicorn[standard]>=0.24.0
pyyaml>=6.0
